COM-3259 - move file
