+++ tests/unit/helpers/customers.test.js
@@ -6,7 +6,7 @@ const crypto = require('crypto');
 const moment = require('moment');
 const QRCode = require('qrcode');
 const SinonMongoose = require('../sinonMongoose');
-const CustomerQRCode = require('../../../src/data/pdf/customerQRCode/customerQRCode');
+const CustomerQRCode = require('../../../src/data/pdf/customerQRCode');
 const Customer = require('../../../src/models/Customer');
 const CustomerPartner = require('../../../src/models/CustomerPartner');
 const Event = require('../../../src/models/Event');
